requirements.txt
aiohttp
aiosignal
altair
async-timeout
attrs
bcrypt
beautifulsoup4
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
contourpy
cycler
et_xmlfile
exceptiongroup
filelock
fonttools
frozenlist
fsspec
gitdb
GitPython
greenlet
h11
huggingface-hub
idna
Jinja2
joblib
jsonschema
jsonschema-specifications
kiwisolver
magiclog
MarkupSafe
matplotlib
multidict
mysql-connector-python
nest-asyncio
nltk
numpy
openpyxl
outcome
packaging
pandas
pillow
playwright
plotly
protobuf
pyarrow
pycparser
pydeck
pyee
pyparsing
PySocks
python-dateutil
python-dotenv
pytz
PyYAML
referencing
regex
requests
rpds-py
safetensors
selenium
six
smmap
sniffio
sortedcontainers
soupsieve
stackclimber
streamlit
streamlit-aggrid
streamlit-option-menu
tenacity
textblob
tokenizers
toml
tornado
tqdm
transformers
trio
trio-websocket
typing_extensions
tzdata
undetected-chromedriver
urllib3
watchdog
websocket-client
websockets
wsproto
yarl

